+++ lib/noticed/delivery_methods/database.rb
@@ -3,10 +3,14 @@ module Noticed
     class Database < Base
       # Must return the database record
       def deliver
-        recipient.notifications.create(attributes)
+        recipient.send(association_name).create!(attributes)
       end
 
       private
+
+      def association_name
+        options[:association] || :notifications
+      end
 
       def attributes
         if (method = options[:format])