+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Translation helpers
 * Allow notification objects to call Rails url helpers directly
 * Add `noticed:notification` generator
+* Allow changing database association name with `deliver_by :database, association: :notifications`
 
 ### 1.1.0